feat(): add the ability to give posts reactions (likes or dislikes)
Added migration for the table with reactions; reactions under posts or comments of other users will be stored there.
Configured polymorphic relationships for tables.
A service(ReactionService) has been registered that is responsible for setting reactions and displaying the number of likes and dislikes.
An enam class (ReactionType) has been created that is responsible for the reaction type